services/structure/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "structure"
version = "0.1.0"
description = "R&D Orchestration Topology - Deterministic Kernels + LLM Interface"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "httpx>=0.27.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
    "jsonschema>=4.21",
]

[tool.setuptools.packages.find]
include = [
    "gateway*",
    "router*",
    "validator*",
    "kernels*",
    "models*",
    "runtime*",
    "plugins*",
    "telemetry*",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
# F401: unused imports (framework defines interfaces ahead of implementation)
# E402: module-level import not at top (tests need path manipulation)
# F841: unused variable (scaffolding for future implementation)
ignore = ["F401", "E402", "F841"]
select = ["E", "F", "W"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E402"]
